Extends the architecture to support different cameras and adds a new animated spawning point camera in debug mode (see README for more details on that).
- Addresses #196
- Fixes some minor issues related to the rendering of the UI
- Adds new console commands for generating and loading worlds (see #71)
- Updates README to represent the latest changes and features

+++ src/org/terasology/rendering/shader/ShaderParametersGelCube.java
@@ -36,7 +36,7 @@ public class ShaderParametersGelCube implements IShaderParameters {
             program.setInt("carryingTorch", tera.getActivePlayer().isCarryingTorch() ? 1 : 0);
         }
         if (tera.getActiveWorldRenderer() != null) {
-            program.setInt("tick", tera.getActiveWorldRenderer().getTick());
+            program.setFloat("tick", (float) tera.getActiveWorldRenderer().getTick());
         }
     }
